pyproject.toml
[project]
name = "projeto-pedido"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "rumians", email = "ruanfort12@gmail.com"},
]
dependencies = ["django>=4.2.19", "djangorestframework>=3.15.2"]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
